move mapper to generic sync
